sub2/backend/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.9
joblib==0.14.0
numpy==1.17.3
pandas==0.25.3
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
djangorestframework-jwt==1.11.0
django-cors-headers==3.2.1